% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_path.R
\name{abs_path}
\alias{abs_path}
\title{Absolute (normalized) path}
\usage{
abs_path(..., mustWork = FALSE)
}
\arguments{
\item{\dots}{character, as in \link{file.path}}

\item{mustWork}{as in \link{normalizePath}}
}
\value{
character
}
\description{
Absolute (normalized) path
}
