% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{read_attributes}
\alias{read_attributes}
\alias{write_attributes}
\title{Read and write attributes}
\usage{
read_attributes(path)

write_attributes(x, path, exclude = c("names"))
}
\arguments{
\item{path}{Directory}

\item{x}{Object whose attributes to write}

\item{exclude}{Attributes to ignore. Default to `names`, as names are handled
separately.}
}
\description{
Read and write attributes
}
