% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache}
\alias{cache}
\title{Cache R objects on disk}
\usage{
cache(expr, key, verbose = getOption("verbose"), force = FALSE)
}
\arguments{
\item{expr}{expression}

\item{key}{character (filename will be \code{key + .rds})}

\item{verbose}{logical}

\item{force}{logical}

\item{path}{character}

\item{compress}{as in \link{readRDS}}

\item{\dots}{further arguments}
}
\description{
Store and retrieve results of arbitrary R expressions in .Rds files
}
\note{
Watch out for case-insensitive filesystems!
}
