% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheFile.R
\name{cache_file_state_clear}
\alias{cache_file_state_clear}
\title{Clear the in-memory file state cache}
\usage{
cache_file_state_clear()
}
\value{
Invisible integer: number of entries cleared.
}
\description{
Removes all cached file fingerprints and hashes. Subsequent file hash
lookups will re-read file metadata and recompute hashes.
}
