% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_meta.R
\name{get_cache_meta}
\alias{get_cache_meta}
\alias{get_cache_type}
\alias{set_cache_meta}
\alias{update_cache_meta}
\title{Read and write cache meta and type.}
\usage{
get_cache_meta(path)

get_cache_type(path)

set_cache_meta(path, meta)

update_cache_meta(path, ...)
}
\arguments{
\item{path}{Directory}

\item{meta}{Named list to write in metadata file}

\item{...}{Named arguments to be added in meta file}
}
\description{
Read and write cache meta and type.
}
