% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheFile.R
\name{cache_tree_files}
\alias{cache_tree_files}
\title{List All Tracked Files}
\usage{
cache_tree_files()
}
\value{
A character vector of unique file paths.
}
\description{
Returns a sorted character vector of all file paths tracked across
all graph nodes.
}
