% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached.R
\name{cached}
\alias{cached}
\title{Declare a cached object}
\usage{
cached(..., root = cache_root(), ext = c(".rds", ".feather", ".rds.lz4",
  ".rds.snz"), verbose = getOption("verbose"))
}
\arguments{
\item{\dots}{components of cache key (NOT the cached expression!)}

\item{root}{function that yields the path on disk}

\item{ext}{form of cached data (".feather" is fastest for r/w, but larger on disk)}

\item{verbose}{logical}
}
\description{
Declare a cached object
}
\examples{
cached("exp", "1", verbose = TRUE) \%or\% exp(1)

}
\seealso{
lazy_or
}
