% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacher-package.R
\docType{package}
\name{cacher}
\alias{cacher}
\alias{cacher-package}
\title{In-memory cache interface.}
\description{
This package exposes one function: \code{LRUcache}
Please take a look at its documentation or open an issue on
https://github.com/kirillseva/cacher if you have any questions
}

