% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheFile.R, R/cacheTree.R
\name{cacheR_default_dir}
\alias{cacheR_default_dir}
\title{Default Cache Directory}
\usage{
cacheR_default_dir()

cacheR_default_dir()
}
\description{
Default Cache Directory

Get Default Cache Directory
}
