% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheFile.R
\name{cacheTree_for_file}
\alias{cacheTree_for_file}
\title{Find nodes associated with a specific file path}
\usage{
cacheTree_for_file(path)
}
\arguments{
\item{path}{Character path to the file to look up.}
}
\value{
A list of nodes that reference the given file.
}
\description{
Find nodes associated with a specific file path
}
