% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheFile.R
\name{cacheTree_load}
\alias{cacheTree_load}
\title{Load an execution graph from disk}
\usage{
cacheTree_load(path)
}
\arguments{
\item{path}{File path to load from}
}
\description{
Load an execution graph from disk
}
