% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheTree.R
\name{cacheTree_nodes}
\alias{cacheTree_nodes}
\title{Return cache tree nodes}
\usage{
cacheTree_nodes()
}
\value{
An object representing nodes (e.g. a data.frame with
\code{id}, \code{fn_label}, \code{key}, \code{cache_file}, etc.).
}
\description{
Returns a list/data.frame describing all nodes currently recorded
in the cache tree (each cached call instance).
}
