% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheFile.R
\name{cacheTree_save}
\alias{cacheTree_save}
\title{Save the execution graph to disk}
\usage{
cacheTree_save(path)
}
\arguments{
\item{path}{File path to save to}
}
\description{
Save the execution graph to disk
}
