% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheFile.R
\name{cacheTree_sync}
\alias{cacheTree_sync}
\title{Load graph from disk into memory}
\usage{
cacheTree_sync(cache_dir)
}
\arguments{
\item{cache_dir}{Character path to the cache directory containing \code{graph.rds}.}
}
\value{
Invisible \code{NULL}.
}
\description{
Load graph from disk into memory
}
