% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_if_needed}
\alias{create_if_needed}
\title{Create a directory if it does not already exist.}
\usage{
create_if_needed(dir)
}
\arguments{
\item{dir}{Directory to create}
}
\value{
TRUE on success
}
\description{
Create the directory or does nothing it it already exists.
Stops with an error if the directory creation fails.
}
