% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorator}
\alias{decorator}
\title{Turn a function into a decorator.}
\usage{
decorator(f)
}
\arguments{
\item{f}{A function to wrap as a decorator.}
}
\value{
A decorator object (function with class \code{"decorator"}).
}
\description{
Turn a function into a decorator.
}
