% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheFile.R
\name{.append_graph_to_disk}
\alias{.append_graph_to_disk}
\title{Append graph changes to disk (Persistence Layer)}
\usage{
.append_graph_to_disk(cache_dir, new_node = NULL, new_edge = NULL)
}
\description{
Append graph changes to disk (Persistence Layer)
}
\keyword{internal}
