% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{.delayed_decorate}
\alias{.delayed_decorate}
\title{Combine two decorators lazily into a new decorator.}
\usage{
.delayed_decorate(d1, d2, decorator_calls)
}
\description{
Combine two decorators lazily into a new decorator.
}
\keyword{internal}
