% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheFile.R
\docType{data}
\name{.file_state_cache}
\alias{.file_state_cache}
\title{File State Cache}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.file_state_cache
}
\description{
File State Cache
}
\keyword{internal}
