% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheFile.R
\docType{data}
\name{.file_state_cache_limit}
\alias{.file_state_cache_limit}
\title{Maximum entries in .file_state_cache before eviction}
\format{
An object of class \code{integer} of length 1.
}
\usage{
.file_state_cache_limit
}
\description{
Maximum entries in .file_state_cache before eviction
}
\keyword{internal}
