% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheFile.R
\name{.find_path_specs}
\alias{.find_path_specs}
\title{Scan function AST for path-like function calls}
\usage{
.find_path_specs(expr)
}
\description{
Scan function AST for path-like function calls
}
\keyword{internal}
