% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheFile.R
\name{.get_scoped_env_hash}
\alias{.get_scoped_env_hash}
\title{Helper: Hash ONLY the globals used by the function (Scope Isolation)}
\usage{
.get_scoped_env_hash(fun)
}
\description{
Helper: Hash ONLY the globals used by the function (Scope Isolation)
}
\keyword{internal}
