% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheFile.R
\docType{data}
\name{.graph_cache}
\alias{.graph_cache}
\title{Execution Graph Cache}
\format{
An object of class \code{environment} of length 3.
}
\usage{
.graph_cache
}
\description{
Execution Graph Cache
}
\keyword{internal}
