% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheFile.R
\name{.hash_inputs}
\alias{.hash_inputs}
\title{Helper to hash inputs}
\usage{
.hash_inputs(inputs, hash_file_paths, file_pattern = NULL, algo = "xxhash64")
}
\description{
Helper to hash inputs
}
\keyword{internal}
