% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.load_cacheR_config}
\alias{.load_cacheR_config}
\title{Load .cacheR.yml config from working directory}
\usage{
.load_cacheR_config(path = file.path(getwd(), ".cacheR.yml"))
}
\arguments{
\item{path}{Path to the YAML config file.}
}
\value{
Invisible NULL.
}
\description{
Reads a YAML config file and sets options that haven't already been set
by the user. Precedence: explicit function args > R options > .cacheR.yml > defaults.
}
\keyword{internal}
