% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheFile.R
\name{.probabilistic_file_hash}
\alias{.probabilistic_file_hash}
\title{Probabilistic File Hashing}
\usage{
.probabilistic_file_hash(
  path,
  block_size = 64 * 1024,
  n_blocks = 5,
  algo = "xxhash64",
  full_hash_limit = 5 * 1024 * 1024
)
}
\description{
Probabilistic File Hashing
}
\keyword{internal}
