% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheFile.R
\name{.replace_paths_with_hashes}
\alias{.replace_paths_with_hashes}
\title{Substitute Paths with Hashes}
\usage{
.replace_paths_with_hashes(x, file_pattern = NULL, algo = "xxhash64")
}
\description{
Substitute Paths with Hashes
}
\keyword{internal}
