% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eval_text}
\alias{eval_text}
\title{Utility to evalulate text as an expression}
\usage{
eval_text(text, env)
}
\arguments{
\item{text}{a string which can be evaluated as an expression}

\item{env}{an environment}
}
\value{
The result of running \code{text} as an expression in \code{env}
}
\description{
Evaluate text as an expression in a given environment
}
