% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheFile.R
\name{export_targets_file}
\alias{export_targets_file}
\title{Export to targets}
\usage{
export_targets_file(path = "_targets.R")
}
\arguments{
\item{path}{Character path to write the targets pipeline script.}
}
\value{
Invisible \code{NULL}.
}
\description{
Export to targets
}
