% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{get.cache.dir}
\alias{get.cache.dir}
\title{Returns the path of the caching directory path}
\usage{
get.cache.dir()
}
\value{
the path of the caching directory
}
\description{
Returns the path of the caching directory path
}
