% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{get.object.cached}
\alias{get.object.cached}
\title{Return a list of the cached files}
\usage{
get.object.cached()
}
\value{
vector of the names of the cached objects (as strings)
}
\description{
Return a list of the cached files
}
\examples{
get.object.cached()
}
