% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{\%@\%.decorator}
\alias{\%@\%.decorator}
\title{Core method: apply a decorator to a function.
Supports chaining: d1 \%@\% d2 \%@\% function(...) ...}
\usage{
\method{\%@\%}{decorator}(decorator, f)
}
\description{
Core method: apply a decorator to a function.
Supports chaining: d1 \%@\% d2 \%@\% function(...) ...
}
