% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{\%@\%}
\alias{\%@\%}
\title{Decorator application operator}
\usage{
decorator \%@\% f
}
\arguments{
\item{decorator}{A decorator object, such as that returned by \code{\link[=cacheFile]{cacheFile()}}.}

\item{fun}{A function to decorate.}
}
\value{
A new function with caching behavior.
}
\description{
Applies a decorator returned by \code{\link[=cacheFile]{cacheFile()}} to a function.
}
