% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{inspect_cache}
\alias{inspect_cache}
\title{Show all files in a cached object.}
\usage{
inspect_cache(path)
}
\arguments{
\item{path}{Path to a stored object.}
}
\description{
A simple wrapper around list.files.
Useful for debugging.
}
