% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{list.objects}
\alias{list.objects}
\title{List all objects and return them in a dataframe associated with their type}
\usage{
list.objects(env = .GlobalEnv)
}
\value{
dataframe of boject names associated with their types
}
\description{
List all objects and return them in a dataframe associated with their type
}
\examples{
list.objects()
}
