% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{load.all.cached.objects}
\alias{load.all.cached.objects}
\title{Loads all cached objects in the global environment. Used to restore an envirnoment quickly}
\usage{
load.all.cached.objects()
}
\value{
nothing
}
\description{
Loads all cached objects in the global environment. Used to restore an envirnoment quickly
}
\examples{
load.all.cached.objects()
}
