% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{load.object}
\alias{load.object}
\title{Load an object (cache)}
\usage{
load.object(name, global = F)
}
\arguments{
\item{name}{name of the object to load}

\item{global}{load in the global environment (if TRUE) or as an object (if FALSE, default)}
}
\value{
object or nothing
}
\description{
Load an object (cache)
}
\examples{
load.object("iris", global=T)
iris2 <- load.object("iris")
}
