% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plaintext_writer.R
\name{logical2int}
\alias{logical2int}
\alias{int2logical}
\title{Logical / Integer conversion}
\usage{
logical2int(vec)

int2logical(vec)
}
\arguments{
\item{vec}{Logical or integer vector to convert}
}
\description{
(TRUE, FALSE, NA) <-> (1, 0, NA)
}
\section{Functions}{
\itemize{
\item \code{int2logical}: Integer (0, 1, NA) to Logical (FALSE, TRUE, NA)
}}

