% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lru.R
\name{LRUcache}
\alias{LRUcache}
\title{Create an LRU in-memory cache}
\usage{
LRUcache(params)
}
\arguments{
\item{params}{numeric or character. If numeric it will create a cache that's
limited by the total number of objects it can hold. If it's a character it
will check how much RAM it consumes using \code{pryr::object_size}}
}
\description{
LRU stands for Least Recently Used and is a description of a cache eviction
strategy. In simpler words, this means that you can set a size limit
and the cache will smartly forget about the item that was accessed last
in case you need to add another item which would cause the cache to breach the
memory or size constraint.
}
\details{
You can use cacher either by calling \code{LRUcache(3)} to store maximum 3 objects
or calling \code{LRUcache("150mb")} to limit the cache size to not exceed 150mb of RAM.
}

