% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{object.cached}
\alias{object.cached}
\title{Check if an object is cached}
\usage{
object.cached(name)
}
\arguments{
\item{name}{name of the object to search}
}
\value{
bool TRUE if object is cached, else FALSE
}
\description{
Check if an object is cached
}
\examples{
object.cached("iris")
}
