% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_call}
\alias{parse_call}
\title{Parse a fucntion call into it's constituent parts}
\usage{
parse_call(q)
}
\arguments{
\item{q}{function call/assignment of the form f <- g()}
}
\value{
List of: enviroment the call was made and (all as text) the call
itself, name of the variable being assigned to, name of the function, name
of the arguments the function takes
}
\description{
Parse a function call/assignment into text pieces - function name, argument
names and variable name as well as the environment it has been called in
}
