% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_text}
\alias{parse_text}
\title{Parse the text of the function call/assignment into it's constituent parts}
\usage{
parse_text(text)
}
\arguments{
\item{text}{the text of the function call/assignment}
}
\value{
A list of (all as text) the whole call, the function name, the
variable name, the names of the arguments to the function
}
\description{
Parses the text of a function call, breaking it into the constituent parts
needed to test whether the call has changed
}
