% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plaintext_factory.R
\name{plaintext_writer_factory}
\alias{plaintext_writer_factory}
\alias{plaintext_reader_factory}
\title{Plaintext reader and writer factories}
\usage{
plaintext_writer_factory(type, cast = function(x) x,
  cast_args = list())

plaintext_reader_factory(cast = function(x) x)
}
\arguments{
\item{type}{cache_type to write}

\item{cast}{Optional cast function when reading or writing.}

\item{cast_args}{Optional extra arguments to the cast function.}
}
\description{
Plaintext reader and writer factories
}
