% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plaintext_writer.R
\name{plaintext_writer}
\alias{plaintext_writer}
\alias{plaintext_reader}
\title{Plaintext reader and writer}
\usage{
plaintext_writer(type, cast = function(x) x, cast_args = list())

plaintext_reader(cast = function(x) x)
}
\arguments{
\item{type}{cache_type to write}

\item{cast}{Optional cast function when reading or writing.}

\item{cast_args}{Optional extra arguments to the cast function.}
}
\description{
Plaintext reader and writer
}
