% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{pretty_code}
\alias{pretty_code}
\title{Get a "pretty" code representation for a function.}
\usage{
pretty_code(f)
}
\arguments{
\item{f}{A decorated function.}
}
\value{
The source reference or body of the function.
}
\description{
Get a "pretty" code representation for a function.
}
