% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{print.decorated}
\alias{print.decorated}
\title{Pretty-print decorated functions.}
\usage{
\method{print}{decorated}(x, useSource = TRUE, ...)
}
\arguments{
\item{x}{A decorated function.}

\item{useSource}{Logical; use source references if available.}

\item{...}{Additional arguments passed to \code{print.function}.}
}
\value{
Invisible \code{x}.
}
\description{
Pretty-print decorated functions.
}
