% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cache.R
\name{read_cache}
\alias{read_cache}
\title{Read data previously saved with write_cache}
\usage{
read_cache(path)
}
\arguments{
\item{path}{Path (including object name)}
}
\description{
Read data previously saved with write_cache
}
\seealso{
\link{write_cache}
}
