% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cache.R
\name{read_cache.rds.gz}
\alias{read_cache.rds.gz}
\alias{read_cache.list}
\title{Internal read_cache functions.}
\usage{
read_cache.rds.gz(path)

read_cache.list(path)
}
\arguments{
\item{path}{Path (including object name)}
}
\description{
This functions are called by `read_cache` when appropriate.
}
\seealso{
read_cache
}
