% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cache.R
\name{read_cache_recursive}
\alias{read_cache_recursive}
\alias{read_cache_recursive.list}
\alias{read_cache_recursive.data.frame}
\alias{read_cache_recursive.character}
\alias{read_cache_recursive.factor}
\alias{read_cache_recursive.numeric}
\alias{read_cache_recursive.integer}
\alias{read_cache_recursive.logical}
\alias{read_cache_recursive.complex}
\title{Internal read_cache functions.}
\usage{
read_cache_recursive(name, path)

read_cache_recursive.list(path)

read_cache_recursive.data.frame(path)

read_cache_recursive.character(path)

read_cache_recursive.factor(path)

read_cache_recursive.numeric(path)

read_cache_recursive.integer(path)

read_cache_recursive.logical(path)

read_cache_recursive.complex(path)
}
\arguments{
\item{name}{Name of sub-directory}

\item{path}{Root path}
}
\description{
`read_cache_recursive` reads a cache without any checks. Always used internally.
`read_cache` is the user-facing entry point, performing sense checks.
}
\seealso{
read_cache
}
