% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{save.all.objects}
\alias{save.all.objects}
\title{Saves all cached objects in the global environment. Used to save an envirnoment quickly.
This seems to be much faster than save.data It is also much more modular. It alows to
load individual objects in other projects for example to combine data analysed separately.}
\usage{
save.all.objects(env)
}
\value{
nothing
}
\description{
Saves all cached objects in the global environment. Used to save an envirnoment quickly.
This seems to be much faster than save.data It is also much more modular. It alows to
load individual objects in other projects for example to combine data analysed separately.
}
\examples{
save.all.objects()
}
