% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{save.object}
\alias{save.object}
\title{Save an object (cache)}
\usage{
save.object(o, name = NULL, cache.dir)
}
\arguments{
\item{o}{object to save}
}
\value{
nothing
}
\description{
Save an object (cache)
}
\examples{
save.object(iris)
}
