% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{set.cache.dir}
\alias{set.cache.dir}
\title{Sets the caching directory path}
\usage{
set.cache.dir(path)
}
\arguments{
\item{path}{the path of the caching directory}
}
\description{
Sets the caching directory path
}
