% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cache.R
\name{set_cache_meta}
\alias{set_cache_meta}
\alias{set_cache_type}
\title{Write an appropriate .cache_meta to a directory}
\usage{
set_cache_meta(path, meta)

set_cache_type(path, type)
}
\arguments{
\item{path}{Directory}

\item{meta}{Metadata object (named list)}

\item{type}{Cache type (set as `cache_type` field in `.cache_meta`)}
}
\description{
Write an appropriate .cache_meta to a directory
}
\section{Functions}{
\itemize{
\item \code{set_cache_type}: Shortcut to set cache_type
}}

