\name{showfiles}
\alias{showfiles}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Show source files}
\description{
  Show the source files that are available for examination
}
\usage{
showfiles()
}
%- maybe also 'usage' for other objects documented here.
\details{
  This function can be used to show what source files are available in
  the cache to examined by the user.  One can switch between different
  source files by calling the \code{sourcefile} function.
}
\value{
  A character vector of the available source files.
}
\seealso{\code{\link{sourcefile}}}
\keyword{utilities}
