% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_split_trim}
\alias{str_split_trim}
\title{Utility to split a string and trim all whitespace}
\usage{
str_split_trim(string, pattern, n = Inf)
}
\arguments{
\item{string}{a string}

\item{pattern}{a string which string can be split on}

\item{n}{number of substring string should be split into}
}
\value{
A character vector of the substring of \code{string} split on
\code{pattern} with all whitespace removed (from both sides)
}
\description{
Split a string and trim all the resulting substrings
}
