% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{uncache.object}
\alias{uncache.object}
\title{Remove object from cache}
\usage{
uncache.object(name)
}
\value{
name of the object removed
}
\description{
Remove object from cache
}
\examples{
uncache.object('df')
}
