% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cache.R
\name{write_cache}
\alias{write_cache}
\title{Write a named list to cache}
\usage{
write_cache(x, path, name = NULL, overwrite = FALSE, ...)
}
\arguments{
\item{x}{Object to save.}

\item{path}{Cache directory. A subdirectory will be created.}

\item{name}{Optional: name of the directory to create. Default = NULL to use
the expression passed as `x`.}

\item{overwrite}{Set to TRUE to overwrite an existing cache or directory.}

\item{...}{Passed to write functions.}
}
\description{
Write a named list to cache
}
