% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cache.R
\name{write_cache_recursive}
\alias{write_cache_recursive}
\alias{write_cache_recursive.default}
\alias{write_cache_recursive.list}
\alias{write_cache_recursive.data.frame}
\title{Generic for write_cache_recursive}
\usage{
write_cache_recursive(x, path, ...)

\method{write_cache_recursive}{default}(x, path, ...)

\method{write_cache_recursive}{list}(x, path, ...)

\method{write_cache_recursive}{data.frame}(x, path, ...)
}
\arguments{
\item{x}{An object to write}

\item{path}{A path to write it to}

\item{...}{Pasedd to write functions}
}
\description{
Internal workhorse for write_cache.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method (save as rds.gz)

\item \code{list}: List method (save as subdirectories)

\item \code{data.frame}: Data.frame method (named list with attributes)
}}

