% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R, R/cache_get.R, R/cache_put.R
\name{cache}
\alias{cache}
\alias{cache_get}
\alias{cache_put}
\title{Cache R objects on disk}
\usage{
cache(expr, key, verbose = getOption("verbose"), force = FALSE)

cache_get(
  ...,
  root = cache_root(),
  ext = c(".rds", ".fst", ".feather", ".rds.lz4", ".rds.snz"),
  verbose = getOption("verbose")
)

cache_put(
  obj,
  ...,
  root = cache_root(),
  ext = c(".rds", ".feather", ".rds.lz4", ".rds.snz"),
  verbose = getOption("verbose")
)
}
\arguments{
\item{expr}{expression}

\item{key}{character (filename will be equal to \code{key} appended with \code{ext})}

\item{verbose}{display messages}

\item{force}{ignore whether an object may already be cached}

\item{obj}{object}

\item{compress}{as in \code{\link[readr:read_rds]{readr::read_rds()}}}
}
\value{
object (see Details)
}
\description{
Store and retrieve results of arbitrary R expressions
}
\details{
If \code{ext} is ".fst" or ".feather", then the returned value will be
tabular data. If \code{ext} is ".rds", then it can be any kind of object.
}
