% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_meta.R
\name{get_cache_meta}
\alias{get_cache_meta}
\alias{get_cache_type}
\alias{set_cache_meta}
\alias{set_cache_type}
\title{Read and write cache meta and type.}
\usage{
get_cache_meta(path)

get_cache_type(path)

set_cache_meta(path, meta)

set_cache_type(path, type)
}
\arguments{
\item{meta}{Metadata object (named list)}

\item{type}{Cache type (set as `cache_type` field in `.cache_meta`)}

\item{Path}{Directory}
}
\description{
Read and write cache meta and type.
}
