% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheFile.R
\name{cacheFile}
\alias{cacheFile}
\title{Create a caching decorator}
\usage{
cacheFile(
  cache_dir = NULL,
  backend = getOption("cacheR.backend", "rds"),
  ignore_args = NULL,
  file_pattern = NULL,
  env_vars = NULL,
  hash_file_paths = TRUE,
  algo = "xxhash64"
)
}
\description{
Create a caching decorator
}
