% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheFile.R, R/cacheTree.R
\name{cachePrune}
\alias{cachePrune}
\title{Prune old cache files}
\usage{
cachePrune(cache_dir = cacheR_default_dir(), days_old = 30)

cachePrune(cache_dir = cacheR_default_dir(), days_old = 30)
}
\arguments{
\item{cache_dir}{Directory to prune}

\item{days_old}{Delete files not accessed/modified in X days}
}
\description{
Prune old cache files

Prune old cache files
}
