% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheR-package.R
\docType{package}
\name{cacheR-package}
\alias{cacheR}
\alias{cacheR-package}
\title{cacheR: git-friendly disk caching}
\description{
I/O functions to read/write named lists to/from disk, 
  using cvs-friendly plaintext when possible.
}
\author{
\strong{Maintainer}: Antoine Sachet \email{antoine@soundout.com}

}
\keyword{internal}
