% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheR-package.R
\docType{package}
\name{cacheR-package}
\alias{cacheR}
\alias{cacheR-package}
\title{cacheR: Save Data in Git-friendly Format}
\description{
Read and write any R object to disk using plaintext as much as possible.
}
\examples{
library("cacheR")
path <- tempdir()

my_data <-
  list(data = mtcars,
       model = lm(mpg ~ gear, data = mtcars),
       details = list(date = "2030-01-01",
                      version = "1.2"))

write_cache(my_data, path)

cache <- read_cache("my_data", path)

all.equal(my_data, cache)
# TRUE, of course!

unlink(path)
}
\seealso{
\link{write_cache} \link{read_cache}
}
\author{
\strong{Maintainer}: Antoine Sachet \email{antoine.sac@gmail.com}

}
\keyword{internal}
