% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheR-package.R
\docType{package}
\name{cacheR-package}
\alias{cacheR}
\alias{cacheR-package}
\title{cacheR: git-friendly disk caching}
\description{
I/O functions to read/write named lists to/from disk, 
  using cvs-friendly plaintext when possible.
}
\examples{
library("cacheR")
path <- tempdir()

my_data <-
  list(data = mtcars,
       model = lm(mpg ~ gear, data = mtcars),
       details = list(date = "2030-01-01",
                      version = "1.2"))

write_cache(my_data, path)

cache <- read_cache(file.path(path, "my_data"))

all.equal(my_data, cache)
# TRUE, of course!

unlink(path)
}
\seealso{
\link{write_cache} \link{read_cache}
}
\author{
\strong{Maintainer}: Antoine Sachet \email{antoine@soundout.com}

}
\keyword{internal}
