% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{grapes-at-grapes}
\alias{grapes-at-grapes}
\alias{\%@\%}
\alias{\%@\%.default}
\alias{\%@\%.decorator}
\title{Decorator application operator}
\usage{
decorator \%@\% f

\method{\%@\%}{default}(decorator, f)

\method{\%@\%}{decorator}(decorator, f)
}
\arguments{
\item{decorator}{A decorator object, such as that returned by \code{\link[=cacheFile]{cacheFile()}}.}

\item{f}{A function to decorate.}
}
\value{
A new function with caching behavior.
}
\description{
Applies a decorator returned by \code{\link[=cacheFile]{cacheFile()}} to a function.
}
