% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cache.R
\name{read_cache}
\alias{read_cache}
\title{Read data previously saved with write_cache}
\usage{
read_cache(name, path)
}
\arguments{
\item{name}{Name of the object to read}

\item{path}{Path to the cache directory}
}
\description{
Read data previously saved with write_cache
}
\seealso{
\link{write_cache}
}
