% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cache.R
\name{write_cache_recursive}
\alias{write_cache_recursive}
\alias{write_cache_recursive.default}
\alias{write_cache_recursive.list}
\alias{write_cache_recursive.data.frame}
\alias{write_cache_recursive.character}
\alias{write_cache_recursive.factor}
\alias{write_cache_recursive.numeric}
\alias{write_cache_recursive.integer}
\alias{write_cache_recursive.logical}
\alias{write_cache_recursive.complex}
\title{Generic for write_cache_recursive}
\usage{
write_cache_recursive(x, path, ...)

\method{write_cache_recursive}{default}(x, path, ...)

\method{write_cache_recursive}{list}(x, path, ...)

\method{write_cache_recursive}{data.frame}(x, path, ...)

\method{write_cache_recursive}{character}(x, path, ...)

\method{write_cache_recursive}{factor}(x, path, ...)

\method{write_cache_recursive}{numeric}(x, path, ...)

\method{write_cache_recursive}{integer}(x, path, ...)

\method{write_cache_recursive}{logical}(x, path, ...)

\method{write_cache_recursive}{complex}(x, path, ...)
}
\arguments{
\item{x}{An object to write}

\item{path}{A path to write it to}

\item{...}{Pasedd to write functions}
}
\description{
Internal workhorse for write_cache.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method (save as rds.gz)

\item \code{list}: List method (save as subdirectories)

\item \code{data.frame}: Data.frame method (named list with attributes)

\item \code{character}: Character method (plaintext)

\item \code{factor}: Factor method (plaintext)

\item \code{numeric}: Numeric method (plaintext)

\item \code{integer}: Integer method (plaintext)

\item \code{logical}: Logical method (plaintext)

\item \code{complex}: Complex method (plaintext)
}}

